pyproject.toml
[project]
name = "linkedin-scraper-mcp"
version = "3.0.2"
description = "MCP server for LinkedIn profile, company, and job scraping with Claude AI integration. Supports direct profile/company/job URL scraping with secure credential storage."
readme = "README.md"
requires-python = ">=3.12"
authors = [
    { name = "Daniel Sticker", email = "daniel@sticker.name" }
]
license = "Apache-2.0"
keywords = [
    "linkedin",
    "mcp",
    "model-context-protocol",
    "scraper",
    "ai",
    "automation",
    "llm",
    "anthropic",
    "claude",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Internet :: WWW/HTTP :: Dynamic Content",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Environment :: Console",
    "Operating System :: OS Independent",
]
dependencies = [
    "fastmcp>=2.14.0",
    "httpx>=0.27.0",
    "inquirer>=3.4.0",
    "linkedin-scraper-patchright>=3.1.1",
    "patchright>=1.40.0",
    "pyperclip>=1.9.0",
    "python-dotenv>=1.1.1",
]

[project.urls]
Homepage = "https://github.com/stickerdaniel/linkedin-mcp-server"
Documentation = "https://github.com/stickerdaniel/linkedin-mcp-server#readme"
Repository = "https://github.com/stickerdaniel/linkedin-mcp-server"
Issues = "https://github.com/stickerdaniel/linkedin-mcp-server/issues"
Changelog = "https://github.com/stickerdaniel/linkedin-mcp-server/releases"

[project.scripts]
linkedin-mcp-server = "linkedin_mcp_server.cli_main:main"
linkedin-scraper-mcp = "linkedin_mcp_server.cli_main:main"
ghl-mcp-server = "ghl_mcp_server.cli_main:main"

[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["linkedin_mcp_server*", "ghl_mcp_server*"]
exclude = ["assets*", "docs*", "tests*"]

[tool.setuptools.package-data]
linkedin_mcp_server = ["py.typed"]

[dependency-groups]
dev = [
    "aiohttp>=3.12.13",
    "pre-commit>=4.2.0",
    "pytest>=8.3.5",
    "pytest-asyncio>=1.0.0",
    "pytest-cov>=6.1.1",
    "pytest-xdist>=3.8.0",
    "ruff>=0.11.11",
    "ty>=0.0.1a12",
]
